% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axioma_equity_optimization_parameters.R
\docType{class}
\name{AxiomaEquityOptimizationParameters}
\alias{AxiomaEquityOptimizationParameters}
\title{AxiomaEquityOptimizationParameters}
\format{
An \code{R6Class} generator object
}
\description{
AxiomaEquityOptimizationParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{strategy}}{\link{Strategy}}

\item{\code{account}}{\link{Account} [optional]}

\item{\code{optimization}}{\link{Optimization} [optional]}

\item{\code{outputtypes}}{\link{OutputTypes}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AxiomaEquityOptimizationParameters$new()}}
\item \href{#method-toJSON}{\code{AxiomaEquityOptimizationParameters$toJSON()}}
\item \href{#method-fromJSON}{\code{AxiomaEquityOptimizationParameters$fromJSON()}}
\item \href{#method-toJSONString}{\code{AxiomaEquityOptimizationParameters$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AxiomaEquityOptimizationParameters$fromJSONString()}}
\item \href{#method-clone}{\code{AxiomaEquityOptimizationParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParameters$new(
  strategy,
  outputtypes,
  account = NULL,
  optimization = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParameters$fromJSON(
  AxiomaEquityOptimizationParametersJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParameters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParameters$fromJSONString(
  AxiomaEquityOptimizationParametersJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
