% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axioma_equity_optimization_parameters_root.R
\docType{class}
\name{AxiomaEquityOptimizationParametersRoot}
\alias{AxiomaEquityOptimizationParametersRoot}
\title{AxiomaEquityOptimizationParametersRoot}
\format{
An \code{R6Class} generator object
}
\description{
AxiomaEquityOptimizationParametersRoot Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{AxiomaEquityOptimizationParameters} [optional]}

\item{\code{meta}}{\link{object} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AxiomaEquityOptimizationParametersRoot$new()}}
\item \href{#method-toJSON}{\code{AxiomaEquityOptimizationParametersRoot$toJSON()}}
\item \href{#method-fromJSON}{\code{AxiomaEquityOptimizationParametersRoot$fromJSON()}}
\item \href{#method-toJSONString}{\code{AxiomaEquityOptimizationParametersRoot$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AxiomaEquityOptimizationParametersRoot$fromJSONString()}}
\item \href{#method-clone}{\code{AxiomaEquityOptimizationParametersRoot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParametersRoot$new(data = NULL, meta = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParametersRoot$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParametersRoot$fromJSON(
  AxiomaEquityOptimizationParametersRootJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParametersRoot$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParametersRoot$fromJSONString(
  AxiomaEquityOptimizationParametersRootJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AxiomaEquityOptimizationParametersRoot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
