% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks_api.R
\docType{class}
\name{BenchmarksApi}
\alias{BenchmarksApi}
\title{Benchmarks operations}
\format{
An \code{R6Class} generator object
}
\description{
factset.analyticsapi.engines.Benchmarks
}
\section{Methods}{

\describe{
\strong{ GetSPARBenchmarkById } \emph{ Get SPAR benchmark details }
This endpoint returns the details of a given SPAR benchmark identifier.

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{SPARBenchmarkRoot} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Expected response, returns the list of prefix and return types for the benchmark.

\item return type : SPARBenchmarkRoot 
\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-FactSet-Api-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-FactSet-Api-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-FactSet-Api-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 400 | Invalid benchmark identifier.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-FactSet-Api-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-FactSet-Api-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-FactSet-Api-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 401 | Missing or invalid authentication.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 403 | User is forbidden with current credentials


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-FactSet-Api-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-FactSet-Api-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-FactSet-Api-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 404 | Benchmark identifier not found.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-FactSet-Api-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-FactSet-Api-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-FactSet-Api-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 406 | Unsupported Accept header. Header needs to be set to application/json.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-FactSet-Api-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-FactSet-Api-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-FactSet-Api-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 429 | Rate limit reached. Wait till the time specified in Retry-After header value to make further requests.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
Retry-After \tab Time to wait in seconds before making a new request as the rate limit has reached. \cr
}
\item status code : 500 | Server error. Log the X-DataDirect-Request-Key header to assist in troubleshooting.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 503 | Request timed out. Retry the request in sometime.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
}

}
}

\examples{
\dontrun{
####################  GetSPARBenchmarkById  ####################

library(factset.analyticsapi.engines)
var.id <- 'id_example' # character | Benchmark Identifier

#Get SPAR benchmark details
api.instance <- BenchmarksApi$new()

#Configure HTTP basic authorization: Basic
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- tryCatch(
            api.instance$GetSPARBenchmarkById(var.id),
            ApiException = function(ex) ex 
         )
# In case of error, print the error object 
if(!is.null(result$ApiException)) {
  cat(result$ApiException$toString())
} else {
# deserialized response object
response.object <- result$content
# response headers
response.headers <- result$response$headers
# response status code
response.status.code <- result$response$status_code
}


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BenchmarksApi$new()}}
\item \href{#method-GetSPARBenchmarkById}{\code{BenchmarksApi$GetSPARBenchmarkById()}}
\item \href{#method-GetSPARBenchmarkByIdWithHttpInfo}{\code{BenchmarksApi$GetSPARBenchmarkByIdWithHttpInfo()}}
\item \href{#method-clone}{\code{BenchmarksApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarksApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSPARBenchmarkById"></a>}}
\if{latex}{\out{\hypertarget{method-GetSPARBenchmarkById}{}}}
\subsection{Method \code{GetSPARBenchmarkById()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarksApi$GetSPARBenchmarkById(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSPARBenchmarkByIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetSPARBenchmarkByIdWithHttpInfo}{}}}
\subsection{Method \code{GetSPARBenchmarkByIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarksApi$GetSPARBenchmarkByIdWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarksApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
