% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiab_calculation_parameters.R
\docType{class}
\name{FIABCalculationParameters}
\alias{FIABCalculationParameters}
\title{FIABCalculationParameters}
\format{
An \code{R6Class} generator object
}
\description{
FIABCalculationParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fiabdocument}}{character [optional]}

\item{\code{account}}{\link{FIABIdentifier}}

\item{\code{dates}}{\link{FIABDateParameters}}

\item{\code{msl}}{character [optional]}

\item{\code{fisettingsdocument}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FIABCalculationParameters$new()}}
\item \href{#method-toJSON}{\code{FIABCalculationParameters$toJSON()}}
\item \href{#method-fromJSON}{\code{FIABCalculationParameters$fromJSON()}}
\item \href{#method-toJSONString}{\code{FIABCalculationParameters$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FIABCalculationParameters$fromJSONString()}}
\item \href{#method-clone}{\code{FIABCalculationParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationParameters$new(
  account,
  dates,
  fiabdocument = NULL,
  msl = NULL,
  fisettingsdocument = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationParameters$fromJSON(FIABCalculationParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationParameters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationParameters$fromJSONString(FIABCalculationParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
