% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiab_calculation_status.R
\docType{class}
\name{FIABCalculationStatus}
\alias{FIABCalculationStatus}
\title{FIABCalculationStatus}
\format{
An \code{R6Class} generator object
}
\description{
FIABCalculationStatus Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{status}}{character [optional]}

\item{\code{startdatetime}}{character [optional]}

\item{\code{completiondatetime}}{character [optional]}

\item{\code{progress}}{integer [optional]}

\item{\code{batchevents}}{list( \link{EventSummary} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FIABCalculationStatus$new()}}
\item \href{#method-toJSON}{\code{FIABCalculationStatus$toJSON()}}
\item \href{#method-fromJSON}{\code{FIABCalculationStatus$fromJSON()}}
\item \href{#method-toJSONString}{\code{FIABCalculationStatus$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FIABCalculationStatus$fromJSONString()}}
\item \href{#method-clone}{\code{FIABCalculationStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatus$new(
  id = NULL,
  status = NULL,
  startdatetime = NULL,
  completiondatetime = NULL,
  progress = NULL,
  batchevents = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatus$fromJSON(FIABCalculationStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatus$fromJSONString(FIABCalculationStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
