% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiab_calculation_status_summary.R
\docType{class}
\name{FIABCalculationStatusSummary}
\alias{FIABCalculationStatusSummary}
\title{FIABCalculationStatusSummary}
\format{
An \code{R6Class} generator object
}
\description{
FIABCalculationStatusSummary Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{requesttime}}{character [optional]}

\item{\code{lastpolltime}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FIABCalculationStatusSummary$new()}}
\item \href{#method-toJSON}{\code{FIABCalculationStatusSummary$toJSON()}}
\item \href{#method-fromJSON}{\code{FIABCalculationStatusSummary$fromJSON()}}
\item \href{#method-toJSONString}{\code{FIABCalculationStatusSummary$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FIABCalculationStatusSummary$fromJSONString()}}
\item \href{#method-clone}{\code{FIABCalculationStatusSummary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatusSummary$new(requesttime = NULL, lastpolltime = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatusSummary$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatusSummary$fromJSON(FIABCalculationStatusSummaryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatusSummary$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatusSummary$fromJSONString(FIABCalculationStatusSummaryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FIABCalculationStatusSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
