% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_portfolio.R
\docType{class}
\name{OptimalPortfolio}
\alias{OptimalPortfolio}
\title{OptimalPortfolio}
\format{
An \code{R6Class} generator object
}
\description{
OptimalPortfolio Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{acctName}}{character [optional]}

\item{\code{excludeZero}}{character [optional]}

\item{\code{archiveDate}}{character [optional]}

\item{\code{ifAcctExists}}{character [optional]}

\item{\code{ifOfdbDateExists}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimalPortfolio$new()}}
\item \href{#method-toJSON}{\code{OptimalPortfolio$toJSON()}}
\item \href{#method-fromJSON}{\code{OptimalPortfolio$fromJSON()}}
\item \href{#method-toJSONString}{\code{OptimalPortfolio$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OptimalPortfolio$fromJSONString()}}
\item \href{#method-clone}{\code{OptimalPortfolio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimalPortfolio$new(
  acctName = NULL,
  excludeZero = NULL,
  archiveDate = NULL,
  ifAcctExists = NULL,
  ifOfdbDateExists = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimalPortfolio$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimalPortfolio$fromJSON(OptimalPortfolioJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimalPortfolio$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimalPortfolio$fromJSONString(OptimalPortfolioJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimalPortfolio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
