% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizations_api.R
\docType{class}
\name{OptimizationsApi}
\alias{OptimizationsApi}
\title{Optimizations operations}
\format{
An \code{R6Class} generator object
}
\description{
factset.analyticsapi.engines.Optimizations
}
\section{Methods}{

\describe{
\strong{ CancelAxiomaOptimizationById } \emph{ Cancel Axioma optimization by id }
This is the endpoint to cancel a previously submitted optimization. Instead of doing a GET on the polling URL, cancel the request by doing a DELETE.

\itemize{
\item \emph{ @param } id character

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 204 | Expected response, optimization was canceled successfully.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 400 | Invalid identifier provided.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 401 | Missing or invalid authentication.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 403 | User is forbidden with current credentials


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 404 | There was no request for the optimization identifier provided, or the request was already canceled for the provided identifier.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 500 | Server error. Log the X-DataDirect-Request-Key header to assist in troubleshooting.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 503 | Request timed out. Retry the request in sometime.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
}

\strong{ GetAxiomaOptimizationById } \emph{ Get Axioma optimization by id }
This is the endpoint to check on the progress of a previously requested optimization.  If the optimization has finished computing, the body of the response will contain result in JSON.  Otherwise, the optimization is still running and the X-FactSet-Api-PickUp-Progress header will contain a progress percentage.

\itemize{
\item \emph{ @param } id character

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Expected response once optimization is completed, returns JSON in the format specified in the Run Optimization endpoint.


\item response headers :

\tabular{ll}{
Content-Encoding \tab Standard HTTP header. Header value based on Accept-Encoding Request header. \cr
Content-Type \tab Standard HTTP header. \cr
Transfer-Encoding \tab Standard HTTP header. Header value will be set to Chunked if Accept-Encoding header is specified. \cr
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 202 | Expected response returned if the optimization is not yet completed, should contain X-FactSet-Api-PickUp-Progress header.


\item response headers :

\tabular{ll}{
X-FactSet-Api-PickUp-Progress \tab FactSet’s progress header. \cr
Cache-Control \tab Standard HTTP header. Header will specify max-age in seconds. Polling can be adjusted based on the max-age value. \cr
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 400 | Invalid identifier provided.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 401 | Missing or invalid authentication.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 403 | User is forbidden with current credentials


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 404 | Optimization result was already returned, provided id was not a requested optimization, or the optimization was cancelled


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 500 | Server error. Log the X-DataDirect-Request-Key header to assist in troubleshooting


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 503 | Request timed out. Retry the request in sometime.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
}

\strong{ RunAxiomaOptimization } \emph{ Run Axioma optimization }
This endpoint runs Axioma optimization specified in the POST body parameters.  It must be used first before polling or cancelling endpoints.  A successful response will contain the URL to poll for the result of the optimization.    Remarks:    * Any settings in POST body will act as a one-time override over the settings saved in the strategy document.

\itemize{
\item \emph{ @param } axioma.equity.optimization.parameters \link{AxiomaEquityOptimizationParameters}

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 202 | Expected response, contains the URL in the Location header to check the status of the optimization.


\item response headers :

\tabular{ll}{
Location \tab URL to poll for the resulting optimization \cr
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 201 | Expected response if optimization is completed within 30 seconds, returns JSON in the format specified in the Run optimization endpoint.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 400 | Invalid POST body.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 401 | Missing or invalid authentication.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 403 | User is forbidden with current credentials


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 404 | One or more optimization settings were unavailable.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 415 | Missing/Invalid Content-Type header. Header needs to be set to application/json.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 429 | Rate limit reached. Cancel older requests using Cancel optimization endpoint or wait for older requests to finish/expire.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
Retry-After \tab Time to wait in seconds before making a new request as the rate limit has reached. \cr
}
\item status code : 500 | Server error. Log the X-DataDirect-Request-Key header to assist in troubleshooting.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 503 | Request timed out. Retry the request in sometime.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet’s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
}

}
}

\examples{
\dontrun{
####################  CancelAxiomaOptimizationById  ####################

library(factset.analyticsapi.engines)
var.id <- 'id_example' # character | from url, provided from the location header in the Run Optimization endpoint

#Cancel Axioma optimization by id
api.instance <- OptimizationsApi$new()

#Configure HTTP basic authorization: Basic
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- tryCatch(
            api.instance$CancelAxiomaOptimizationById(var.id),
            ApiException = function(ex) ex 
         )
# In case of error, print the error object 
if(!is.null(result$ApiException)) {
  cat(result$ApiException$toString())
} else {
# response headers
response.headers <- result$response$headers
# response status code
response.status.code <- result$response$status_code
}


####################  GetAxiomaOptimizationById  ####################

library(factset.analyticsapi.engines)
var.id <- 'id_example' # character | from url, provided from the location header in the Run Optimization endpoint

#Get Axioma optimization by id
api.instance <- OptimizationsApi$new()

#Configure HTTP basic authorization: Basic
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- tryCatch(
            api.instance$GetAxiomaOptimizationById(var.id),
            ApiException = function(ex) ex 
         )
# In case of error, print the error object 
if(!is.null(result$ApiException)) {
  cat(result$ApiException$toString())
} else {
# response headers
response.headers <- result$response$headers
# response status code
response.status.code <- result$response$status_code
}


####################  RunAxiomaOptimization  ####################

library(factset.analyticsapi.engines)
var.axioma.equity.optimization.parameters <- AxiomaEquityOptimizationParameters$new() # AxiomaEquityOptimizationParameters | 

#Run Axioma optimization
api.instance <- OptimizationsApi$new()

#Configure HTTP basic authorization: Basic
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- tryCatch(
            api.instance$RunAxiomaOptimization(axioma.equity.optimization.parameters=var.axioma.equity.optimization.parameters),
            ApiException = function(ex) ex 
         )
# In case of error, print the error object 
if(!is.null(result$ApiException)) {
  cat(result$ApiException$toString())
} else {
# response headers
response.headers <- result$response$headers
# response status code
response.status.code <- result$response$status_code
}


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizationsApi$new()}}
\item \href{#method-CancelAxiomaOptimizationById}{\code{OptimizationsApi$CancelAxiomaOptimizationById()}}
\item \href{#method-CancelAxiomaOptimizationByIdWithHttpInfo}{\code{OptimizationsApi$CancelAxiomaOptimizationByIdWithHttpInfo()}}
\item \href{#method-GetAxiomaOptimizationById}{\code{OptimizationsApi$GetAxiomaOptimizationById()}}
\item \href{#method-GetAxiomaOptimizationByIdWithHttpInfo}{\code{OptimizationsApi$GetAxiomaOptimizationByIdWithHttpInfo()}}
\item \href{#method-RunAxiomaOptimization}{\code{OptimizationsApi$RunAxiomaOptimization()}}
\item \href{#method-RunAxiomaOptimizationWithHttpInfo}{\code{OptimizationsApi$RunAxiomaOptimizationWithHttpInfo()}}
\item \href{#method-clone}{\code{OptimizationsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CancelAxiomaOptimizationById"></a>}}
\if{latex}{\out{\hypertarget{method-CancelAxiomaOptimizationById}{}}}
\subsection{Method \code{CancelAxiomaOptimizationById()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationsApi$CancelAxiomaOptimizationById(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CancelAxiomaOptimizationByIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CancelAxiomaOptimizationByIdWithHttpInfo}{}}}
\subsection{Method \code{CancelAxiomaOptimizationByIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationsApi$CancelAxiomaOptimizationByIdWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAxiomaOptimizationById"></a>}}
\if{latex}{\out{\hypertarget{method-GetAxiomaOptimizationById}{}}}
\subsection{Method \code{GetAxiomaOptimizationById()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationsApi$GetAxiomaOptimizationById(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAxiomaOptimizationByIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetAxiomaOptimizationByIdWithHttpInfo}{}}}
\subsection{Method \code{GetAxiomaOptimizationByIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationsApi$GetAxiomaOptimizationByIdWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunAxiomaOptimization"></a>}}
\if{latex}{\out{\hypertarget{method-RunAxiomaOptimization}{}}}
\subsection{Method \code{RunAxiomaOptimization()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationsApi$RunAxiomaOptimization(
  axioma.equity.optimization.parameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunAxiomaOptimizationWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-RunAxiomaOptimizationWithHttpInfo}{}}}
\subsection{Method \code{RunAxiomaOptimizationWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationsApi$RunAxiomaOptimizationWithHttpInfo(
  axioma.equity.optimization.parameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizationsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
