% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_account_overrides.R
\docType{class}
\name{OptimizerAccountOverrides}
\alias{OptimizerAccountOverrides}
\title{OptimizerAccountOverrides}
\format{
An \code{R6Class} generator object
}
\description{
OptimizerAccountOverrides Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{portfolio}}{character [optional]}

\item{\code{benchmark}}{character [optional]}

\item{\code{riskModelId}}{character [optional]}

\item{\code{currency}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizerAccountOverrides$new()}}
\item \href{#method-toJSON}{\code{OptimizerAccountOverrides$toJSON()}}
\item \href{#method-fromJSON}{\code{OptimizerAccountOverrides$fromJSON()}}
\item \href{#method-toJSONString}{\code{OptimizerAccountOverrides$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OptimizerAccountOverrides$fromJSONString()}}
\item \href{#method-clone}{\code{OptimizerAccountOverrides$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerAccountOverrides$new(
  portfolio = NULL,
  benchmark = NULL,
  riskModelId = NULL,
  currency = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerAccountOverrides$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerAccountOverrides$fromJSON(OptimizerAccountOverridesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerAccountOverrides$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerAccountOverrides$fromJSONString(OptimizerAccountOverridesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerAccountOverrides$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
