% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer_strategy_overrides.R
\docType{class}
\name{OptimizerStrategyOverrides}
\alias{OptimizerStrategyOverrides}
\title{OptimizerStrategyOverrides}
\format{
An \code{R6Class} generator object
}
\description{
OptimizerStrategyOverrides Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tax}}{character [optional]}

\item{\code{objective}}{character [optional]}

\item{\code{constraints}}{named list( character ) [optional]}

\item{\code{alpha}}{character [optional]}

\item{\code{transactionCost}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizerStrategyOverrides$new()}}
\item \href{#method-toJSON}{\code{OptimizerStrategyOverrides$toJSON()}}
\item \href{#method-fromJSON}{\code{OptimizerStrategyOverrides$fromJSON()}}
\item \href{#method-toJSONString}{\code{OptimizerStrategyOverrides$toJSONString()}}
\item \href{#method-fromJSONString}{\code{OptimizerStrategyOverrides$fromJSONString()}}
\item \href{#method-clone}{\code{OptimizerStrategyOverrides$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerStrategyOverrides$new(
  tax = NULL,
  objective = NULL,
  constraints = NULL,
  alpha = NULL,
  transactionCost = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerStrategyOverrides$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerStrategyOverrides$fromJSON(OptimizerStrategyOverridesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerStrategyOverrides$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerStrategyOverrides$fromJSONString(OptimizerStrategyOverridesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerStrategyOverrides$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
