% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_calculation_parameters_root.R
\docType{class}
\name{PubCalculationParametersRoot}
\alias{PubCalculationParametersRoot}
\title{PubCalculationParametersRoot}
\format{
An \code{R6Class} generator object
}
\description{
PubCalculationParametersRoot Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{named list( \link{PubCalculationParameters} ) [optional]}

\item{\code{meta}}{\link{object} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PubCalculationParametersRoot$new()}}
\item \href{#method-toJSON}{\code{PubCalculationParametersRoot$toJSON()}}
\item \href{#method-fromJSON}{\code{PubCalculationParametersRoot$fromJSON()}}
\item \href{#method-toJSONString}{\code{PubCalculationParametersRoot$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PubCalculationParametersRoot$fromJSONString()}}
\item \href{#method-clone}{\code{PubCalculationParametersRoot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PubCalculationParametersRoot$new(data = NULL, meta = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PubCalculationParametersRoot$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PubCalculationParametersRoot$fromJSON(PubCalculationParametersRootJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PubCalculationParametersRoot$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PubCalculationParametersRoot$fromJSONString(PubCalculationParametersRootJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PubCalculationParametersRoot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
