% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_documents_api.R
\docType{class}
\name{StrategyDocumentsApi}
\alias{StrategyDocumentsApi}
\title{StrategyDocuments operations}
\format{
An \code{R6Class} generator object
}
\description{
factset.analyticsapi.engines.StrategyDocuments
}
\section{Methods}{

\describe{
\strong{ GetAxiomaEquityStrategyDocuments } \emph{ Get Axioma Equity strategy documents and sub-directories in a directory }
This endpoint looks up all Axioma Equity strategy documents and sub-directories in a given directory.

\itemize{
\item \emph{ @param } path character
\item \emph{ @returnType } \link{DocumentDirectories} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Expected response, returns a list of Axioma Equity strategy documents and directories

\item return type : DocumentDirectories 
\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 400 | Invalid query parameter or value provided


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 401 | Missing or invalid authentication


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 403 | User is forbidden with current credentials


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 404 | Path not found


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 406 | Unsupported Accept header.  Header needs to be set to application/json


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 429 | Rate limit reached. Wait till the time specified in Retry-After header value to make further requests.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
Retry-After \tab Time to wait in seconds before making a new request as the rate limit has reached. \cr
}
\item status code : 500 | Server error. Log the X-DataDirect-Request-Key header to assist in troubleshooting


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 503 | Request timed out. Retry the request in some time


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
}

\strong{ GetAxiomaFIStrategyDocuments } \emph{ Get Axioma FI strategy documents and sub-directories in a directory }
This endpoint looks up all Axioma FI strategy documents and sub-directories in a given directory.

\itemize{
\item \emph{ @param } path character
\item \emph{ @returnType } \link{DocumentDirectories} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Expected response, returns a list of Axioma FI strategy documents and directories

\item return type : DocumentDirectories 
\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 400 | Invalid query parameter or value provided


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 401 | Missing or invalid authentication


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 403 | User is forbidden with current credentials


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 404 | Path not found


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 406 | Unsupported Accept header.  Header needs to be set to application/json


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 429 | Rate limit reached. Wait till the time specified in Retry-After header value to make further requests.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
Retry-After \tab Time to wait in seconds before making a new request as the rate limit has reached. \cr
}
\item status code : 500 | Server error. Log the X-DataDirect-Request-Key header to assist in troubleshooting


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 503 | Request timed out. Retry the request in some time


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
}

\strong{ GetBarraStrategyDocuments } \emph{ Get Barra strategy documents and sub-directories in a directory }
This endpoint looks up all Barra strategy documents and sub-directories in a given directory.

\itemize{
\item \emph{ @param } path character
\item \emph{ @returnType } \link{DocumentDirectories} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Expected response, returns a list of Barra strategy documents and directories

\item return type : DocumentDirectories 
\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 400 | Invalid query parameter or value provided


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 401 | Missing or invalid authentication


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 403 | User is forbidden with current credentials


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 404 | Path not found


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 406 | Unsupported Accept header.  Header needs to be set to application/json


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 429 | Rate limit reached. Wait till the time specified in Retry-After header value to make further requests.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
Retry-After \tab Time to wait in seconds before making a new request as the rate limit has reached. \cr
}
\item status code : 500 | Server error. Log the X-DataDirect-Request-Key header to assist in troubleshooting


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 503 | Request timed out. Retry the request in some time


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
}

\strong{ GetFPOStrategyDocuments } \emph{ Get FactSet Portfolio Optimizer strategy documents and sub-directories in a directory }
This endpoint looks up all FactSet Portfolio Optimizer strategy documents and sub-directories in a given directory.

\itemize{
\item \emph{ @param } path character
\item \emph{ @returnType } \link{DocumentDirectories} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Expected response, returns a list of FactSet Portfolio Optimizer strategy documents and directories

\item return type : DocumentDirectories 
\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 400 | Invalid query parameter or value provided


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 401 | Missing or invalid authentication


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 403 | User is forbidden with current credentials


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 404 | Path not found


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 406 | Unsupported Accept header.  Header needs to be set to application/json


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 429 | Rate limit reached. Wait till the time specified in Retry-After header value to make further requests.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
Retry-After \tab Time to wait in seconds before making a new request as the rate limit has reached. \cr
}
\item status code : 500 | Server error. Log the X-DataDirect-Request-Key header to assist in troubleshooting


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 503 | Request timed out. Retry the request in some time


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
}

\strong{ GetNorthfieldStrategyDocuments } \emph{ Get Northfield strategy documents and sub-directories in a directory }
This endpoint looks up all Northfield strategy documents and sub-directories in a given directory.

\itemize{
\item \emph{ @param } path character
\item \emph{ @returnType } \link{DocumentDirectories} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Expected response, returns a list of Northfield strategy documents and directories

\item return type : DocumentDirectories 
\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 400 | Invalid query parameter or value provided


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 401 | Missing or invalid authentication


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 403 | User is forbidden with current credentials


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 404 | Path not found


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 406 | Unsupported Accept header.  Header needs to be set to application/json


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
X-RateLimit-Limit \tab Number of allowed requests for the time window. \cr
X-RateLimit-Remaining \tab Number of requests left for the time window. \cr
X-RateLimit-Reset \tab Number of seconds remaining till rate limit resets. \cr
}
\item status code : 429 | Rate limit reached. Wait till the time specified in Retry-After header value to make further requests.


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
Retry-After \tab Time to wait in seconds before making a new request as the rate limit has reached. \cr
}
\item status code : 500 | Server error. Log the X-DataDirect-Request-Key header to assist in troubleshooting


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
\item status code : 503 | Request timed out. Retry the request in some time


\item response headers :

\tabular{ll}{
X-DataDirect-Request-Key \tab FactSet&#39;s request key header. \cr
X-FactSet-Api-Request-Key \tab Key to uniquely identify an Analytics API request. Only available after successful authentication. \cr
}
}

}
}

\examples{
\dontrun{
####################  GetAxiomaEquityStrategyDocuments  ####################

library(factset.analyticsapi.engines)
var.path <- '' # character | The directory to get the strategy documents and sub-directories in

#Get Axioma Equity strategy documents and sub-directories in a directory
api.instance <- StrategyDocumentsApi$new()

#Configure HTTP basic authorization: Basic
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- tryCatch(
            api.instance$GetAxiomaEquityStrategyDocuments(var.path),
            ApiException = function(ex) ex 
         )
# In case of error, print the error object 
if(!is.null(result$ApiException)) {
  cat(result$ApiException$toString())
} else {
# deserialized response object
response.object <- result$content
# response headers
response.headers <- result$response$headers
# response status code
response.status.code <- result$response$status_code
}


####################  GetAxiomaFIStrategyDocuments  ####################

library(factset.analyticsapi.engines)
var.path <- '' # character | The directory to get the strategy documents and sub-directories in

#Get Axioma FI strategy documents and sub-directories in a directory
api.instance <- StrategyDocumentsApi$new()

#Configure HTTP basic authorization: Basic
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- tryCatch(
            api.instance$GetAxiomaFIStrategyDocuments(var.path),
            ApiException = function(ex) ex 
         )
# In case of error, print the error object 
if(!is.null(result$ApiException)) {
  cat(result$ApiException$toString())
} else {
# deserialized response object
response.object <- result$content
# response headers
response.headers <- result$response$headers
# response status code
response.status.code <- result$response$status_code
}


####################  GetBarraStrategyDocuments  ####################

library(factset.analyticsapi.engines)
var.path <- '' # character | The directory to get the strategy documents and sub-directories in

#Get Barra strategy documents and sub-directories in a directory
api.instance <- StrategyDocumentsApi$new()

#Configure HTTP basic authorization: Basic
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- tryCatch(
            api.instance$GetBarraStrategyDocuments(var.path),
            ApiException = function(ex) ex 
         )
# In case of error, print the error object 
if(!is.null(result$ApiException)) {
  cat(result$ApiException$toString())
} else {
# deserialized response object
response.object <- result$content
# response headers
response.headers <- result$response$headers
# response status code
response.status.code <- result$response$status_code
}


####################  GetFPOStrategyDocuments  ####################

library(factset.analyticsapi.engines)
var.path <- '' # character | The directory to get the strategy documents and sub-directories in

#Get FactSet Portfolio Optimizer strategy documents and sub-directories in a directory
api.instance <- StrategyDocumentsApi$new()

#Configure HTTP basic authorization: Basic
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- tryCatch(
            api.instance$GetFPOStrategyDocuments(var.path),
            ApiException = function(ex) ex 
         )
# In case of error, print the error object 
if(!is.null(result$ApiException)) {
  cat(result$ApiException$toString())
} else {
# deserialized response object
response.object <- result$content
# response headers
response.headers <- result$response$headers
# response status code
response.status.code <- result$response$status_code
}


####################  GetNorthfieldStrategyDocuments  ####################

library(factset.analyticsapi.engines)
var.path <- '' # character | The directory to get the strategy documents and sub-directories in

#Get Northfield strategy documents and sub-directories in a directory
api.instance <- StrategyDocumentsApi$new()

#Configure HTTP basic authorization: Basic
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- tryCatch(
            api.instance$GetNorthfieldStrategyDocuments(var.path),
            ApiException = function(ex) ex 
         )
# In case of error, print the error object 
if(!is.null(result$ApiException)) {
  cat(result$ApiException$toString())
} else {
# deserialized response object
response.object <- result$content
# response headers
response.headers <- result$response$headers
# response status code
response.status.code <- result$response$status_code
}


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StrategyDocumentsApi$new()}}
\item \href{#method-GetAxiomaEquityStrategyDocuments}{\code{StrategyDocumentsApi$GetAxiomaEquityStrategyDocuments()}}
\item \href{#method-GetAxiomaEquityStrategyDocumentsWithHttpInfo}{\code{StrategyDocumentsApi$GetAxiomaEquityStrategyDocumentsWithHttpInfo()}}
\item \href{#method-GetAxiomaFIStrategyDocuments}{\code{StrategyDocumentsApi$GetAxiomaFIStrategyDocuments()}}
\item \href{#method-GetAxiomaFIStrategyDocumentsWithHttpInfo}{\code{StrategyDocumentsApi$GetAxiomaFIStrategyDocumentsWithHttpInfo()}}
\item \href{#method-GetBarraStrategyDocuments}{\code{StrategyDocumentsApi$GetBarraStrategyDocuments()}}
\item \href{#method-GetBarraStrategyDocumentsWithHttpInfo}{\code{StrategyDocumentsApi$GetBarraStrategyDocumentsWithHttpInfo()}}
\item \href{#method-GetFPOStrategyDocuments}{\code{StrategyDocumentsApi$GetFPOStrategyDocuments()}}
\item \href{#method-GetFPOStrategyDocumentsWithHttpInfo}{\code{StrategyDocumentsApi$GetFPOStrategyDocumentsWithHttpInfo()}}
\item \href{#method-GetNorthfieldStrategyDocuments}{\code{StrategyDocumentsApi$GetNorthfieldStrategyDocuments()}}
\item \href{#method-GetNorthfieldStrategyDocumentsWithHttpInfo}{\code{StrategyDocumentsApi$GetNorthfieldStrategyDocumentsWithHttpInfo()}}
\item \href{#method-clone}{\code{StrategyDocumentsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAxiomaEquityStrategyDocuments"></a>}}
\if{latex}{\out{\hypertarget{method-GetAxiomaEquityStrategyDocuments}{}}}
\subsection{Method \code{GetAxiomaEquityStrategyDocuments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetAxiomaEquityStrategyDocuments(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAxiomaEquityStrategyDocumentsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetAxiomaEquityStrategyDocumentsWithHttpInfo}{}}}
\subsection{Method \code{GetAxiomaEquityStrategyDocumentsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetAxiomaEquityStrategyDocumentsWithHttpInfo(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAxiomaFIStrategyDocuments"></a>}}
\if{latex}{\out{\hypertarget{method-GetAxiomaFIStrategyDocuments}{}}}
\subsection{Method \code{GetAxiomaFIStrategyDocuments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetAxiomaFIStrategyDocuments(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetAxiomaFIStrategyDocumentsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetAxiomaFIStrategyDocumentsWithHttpInfo}{}}}
\subsection{Method \code{GetAxiomaFIStrategyDocumentsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetAxiomaFIStrategyDocumentsWithHttpInfo(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBarraStrategyDocuments"></a>}}
\if{latex}{\out{\hypertarget{method-GetBarraStrategyDocuments}{}}}
\subsection{Method \code{GetBarraStrategyDocuments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetBarraStrategyDocuments(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBarraStrategyDocumentsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetBarraStrategyDocumentsWithHttpInfo}{}}}
\subsection{Method \code{GetBarraStrategyDocumentsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetBarraStrategyDocumentsWithHttpInfo(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFPOStrategyDocuments"></a>}}
\if{latex}{\out{\hypertarget{method-GetFPOStrategyDocuments}{}}}
\subsection{Method \code{GetFPOStrategyDocuments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetFPOStrategyDocuments(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFPOStrategyDocumentsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetFPOStrategyDocumentsWithHttpInfo}{}}}
\subsection{Method \code{GetFPOStrategyDocumentsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetFPOStrategyDocumentsWithHttpInfo(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNorthfieldStrategyDocuments"></a>}}
\if{latex}{\out{\hypertarget{method-GetNorthfieldStrategyDocuments}{}}}
\subsection{Method \code{GetNorthfieldStrategyDocuments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetNorthfieldStrategyDocuments(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNorthfieldStrategyDocumentsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetNorthfieldStrategyDocumentsWithHttpInfo}{}}}
\subsection{Method \code{GetNorthfieldStrategyDocumentsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$GetNorthfieldStrategyDocumentsWithHttpInfo(path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyDocumentsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
