% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_overrides.R
\docType{class}
\name{StrategyOverrides}
\alias{StrategyOverrides}
\title{StrategyOverrides}
\format{
An \code{R6Class} generator object
}
\description{
StrategyOverrides Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objective}}{character [optional]}

\item{\code{constraints}}{named list( character ) [optional]}

\item{\code{alpha}}{character [optional]}

\item{\code{transactioncost}}{character [optional]}

\item{\code{tax}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StrategyOverrides$new()}}
\item \href{#method-toJSON}{\code{StrategyOverrides$toJSON()}}
\item \href{#method-fromJSON}{\code{StrategyOverrides$fromJSON()}}
\item \href{#method-toJSONString}{\code{StrategyOverrides$toJSONString()}}
\item \href{#method-fromJSONString}{\code{StrategyOverrides$fromJSONString()}}
\item \href{#method-clone}{\code{StrategyOverrides$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyOverrides$new(
  objective = NULL,
  constraints = NULL,
  alpha = NULL,
  transactioncost = NULL,
  tax = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyOverrides$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyOverrides$fromJSON(StrategyOverridesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyOverrides$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyOverrides$fromJSONString(StrategyOverridesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyOverrides$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
