% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_configuration_summary_root.R
\docType{class}
\name{VaultConfigurationSummaryRoot}
\alias{VaultConfigurationSummaryRoot}
\title{VaultConfigurationSummaryRoot}
\format{
An \code{R6Class} generator object
}
\description{
VaultConfigurationSummaryRoot Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{named list( \link{VaultConfigurationSummary} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VaultConfigurationSummaryRoot$new()}}
\item \href{#method-toJSON}{\code{VaultConfigurationSummaryRoot$toJSON()}}
\item \href{#method-fromJSON}{\code{VaultConfigurationSummaryRoot$fromJSON()}}
\item \href{#method-toJSONString}{\code{VaultConfigurationSummaryRoot$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VaultConfigurationSummaryRoot$fromJSONString()}}
\item \href{#method-clone}{\code{VaultConfigurationSummaryRoot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultConfigurationSummaryRoot$new(data = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultConfigurationSummaryRoot$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultConfigurationSummaryRoot$fromJSON(VaultConfigurationSummaryRootJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultConfigurationSummaryRoot$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultConfigurationSummaryRoot$fromJSONString(VaultConfigurationSummaryRootJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VaultConfigurationSummaryRoot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
